+++ src/domain/entities/activity/types/index.ts
@@ -1,6 +1,6 @@
-export * from './activityEntityProps';
-export * from './activityProps';
-export * from './activityType';
+export * from './activity-entity-props';
+export * from './activity-props';
+export * from './activity-type';
 export * from './categories';
-export * from './responseEntityActivity';
+export * from './response-entity-activity';
 export * from './weekly-frequency';
