+++ src/domain/entities/activity/types/response-entity-activity.ts
