+++ src/domain/entities/activity/value-objects/activity-type/activity-type-value-object.spec.ts
